Implementation of new mutation and bee product system

+++ src/main/java/com/emily/apicraft/interfaces/genetics/IBeeModifierProvider.java
@@ -7,7 +7,7 @@ import java.util.function.Function;
 public interface IBeeModifierProvider {
 
     float applyProductivityModifier(float val);
-    int applyLifespanModifier(int val);
+    float applyLifespanModifier(float val);
     float applyMutationModifier(float val);
     Vec3i applyTerritoryModifier(Vec3i val);
     int applyFertilityModifier(int val);
